+++ src/config/constants.ts
@@ -4,3 +4,4 @@ const { width } = Dimensions.get('screen');
 
 export const CARD_WIDTH_FULL = width * 0.85;
 export const IMAGE_IN_LIST = 80;
+export const PROFILE_IMAGE = 200;