+++ src/screens/Home/components/PostCardBody.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
-import { Text } from 'react-native';
 import { IMAGE_IN_LIST } from '~config/constants';
 import styled from 'styled-components/native';
+import { TextStyled } from '~theme/shared';
 
 interface Props {
   body: string;
@@ -14,7 +14,7 @@ const Container = styled.View`
 const PostCardBody = ({ body }: Props) => {
   return (
     <Container>
-      <Text numberOfLines={3}>{body}</Text>
+      <TextStyled numberOfLines={3}>{body}</TextStyled>
     </Container>
   );
 };