+++ src/utils/deviceInfo.ts
@@ -1,0 +1,8 @@
+import { Platform } from 'react-native';
+import DeviceInfo from 'react-native-device-info';
+
+export const isIOS = Platform.OS === 'ios';
+export const hasNotch = DeviceInfo.hasNotch();
+export const isAndroid = !isIOS;
+
+export const isAndroidWithNotch = hasNotch && isAndroid;